+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/components/selectedmessage/SelectedMessageMenu.kt
@@ -111,7 +111,6 @@ private fun SelectedMessageMenuPreview() {
         val messageOptionsStateList = defaultMessageOptionsState(
             selectedMessage = Message(),
             currentUser = User(),
-            isInThread = false,
             ownCapabilities = ChannelCapabilities.toSet(),
         )