+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/util/extensions/MessageOptionItemVisibility.kt
@@ -14,102 +14,100 @@
  * limitations under the License.
  */
 
-@file:Suppress("TooManyFunctions")
+package io.getstream.chat.android.compose.util.extensions
 
-package io.getstream.chat.android.ui.feature.messages.list.internal
-
+import io.getstream.chat.android.compose.ui.components.messageoptions.MessageOptionItemVisibility
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.User
-import io.getstream.chat.android.ui.feature.messages.list.MessageListViewStyle
 
-internal fun MessageListViewStyle.canReplyToMessage(
+internal fun MessageOptionItemVisibility.canReplyToMessage(
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canReplyToMessage(
-    replyEnabled = replyEnabled,
+    replyEnabled = isReplyVisible,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canThreadReplyToMessage(
+internal fun MessageOptionItemVisibility.canThreadReplyToMessage(
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canThreadReplyToMessage(
-    threadsEnabled = threadsEnabled,
+    threadsEnabled = isThreadReplyVisible,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canCopyMessage(
+internal fun MessageOptionItemVisibility.canCopyMessage(
     message: Message,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canCopyMessage(
-    copyTextEnabled = copyTextEnabled,
+    copyTextEnabled = isCopyTextVisible,
     message = message,
 )
 
-internal fun MessageListViewStyle.canEditMessage(
+internal fun MessageOptionItemVisibility.canEditMessage(
     currentUser: User?,
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canEditMessage(
-    editMessageEnabled = editMessageEnabled,
+    editMessageEnabled = isEditMessageVisible,
     currentUser = currentUser,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canDeleteMessage(
+internal fun MessageOptionItemVisibility.canDeleteMessage(
     currentUser: User?,
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canDeleteMessage(
-    deleteMessageEnabled = deleteMessageEnabled,
+    deleteMessageEnabled = isDeleteMessageVisible,
     currentUser = currentUser,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canFlagMessage(
+internal fun MessageOptionItemVisibility.canFlagMessage(
     currentUser: User?,
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canFlagMessage(
-    flagEnabled = flagEnabled,
+    flagEnabled = isFlagMessageVisible,
     currentUser = currentUser,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canPinMessage(
+internal fun MessageOptionItemVisibility.canPinMessage(
     message: Message,
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canPinMessage(
-    pinMessageEnabled = pinMessageEnabled,
+    pinMessageEnabled = isPinMessageVisible,
     message = message,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canBlockUser(
+internal fun MessageOptionItemVisibility.canBlockUser(
     currentUser: User?,
     message: Message,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canBlockUser(
-    blockUserEnabled = blockUserEnabled,
+    blockUserEnabled = isBlockUserVisible,
     currentUser = currentUser,
     message = message,
 )
 
-internal fun MessageListViewStyle.canMarkAsUnread(
+internal fun MessageOptionItemVisibility.canMarkAsUnread(
     ownCapabilities: Set<String>,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canMarkAsUnread(
-    markAsUnreadEnabled = markAsUnreadEnabled,
+    markAsUnreadEnabled = isMarkAsUnreadVisible,
     ownCapabilities = ownCapabilities,
 )
 
-internal fun MessageListViewStyle.canRetryMessage(
+internal fun MessageOptionItemVisibility.canRetryMessage(
     currentUser: User?,
     message: Message,
 ): Boolean = io.getstream.chat.android.ui.common.utils.canRetryMessage(
-    retryMessageEnabled = retryMessageEnabled,
+    retryMessageEnabled = isRetryMessageVisible,
     currentUser = currentUser,
     message = message,
 )